+++ integration-tests/ssr/gatsby-browser.js
@@ -1,0 +1,1 @@
+import "./sample.css"