+++ packages/gatsby-transformer-sharp/src/customize-schema.js
@@ -39,6 +39,7 @@ const {
   WebPOptionsType,
   BlurredOptionsType,
   TransformOptionsType,
+  AVIFOptionsType,
 } = require(`./types`)
 const { stripIndent } = require(`common-tags`)
 const { prefixId, CODES } = require(`./error-utils`)
@@ -410,7 +411,7 @@ const imageNodeType = ({
         type: GraphQLInt,
         description: stripIndent`
         Maximum display width of generated files. 
-        The actual largest image resolution will be this value multipled by the largest value in outputPixelDensities
+        The actual largest image resolution will be this value multiplied by the largest value in outputPixelDensities
         This only applies when layout = FLUID or CONSTRAINED. For other layout types, use "width"`,
       },
       maxHeight: {
@@ -424,7 +425,7 @@ const imageNodeType = ({
         type: GraphQLInt,
         description: stripIndent`
         The display width of the generated image. 
-        The actual largest image resolution will be this value multipled by the largest value in outputPixelDensities
+        The actual largest image resolution will be this value multiplied by the largest value in outputPixelDensities
         Ignored if layout = FLUID or CONSTRAINED, where you should use "maxWidth" instead.
         `,
       },
@@ -479,7 +480,7 @@ const imageNodeType = ({
       },
       quality: {
         type: GraphQLInt,
-        description: `The default quality. This is overriden by any format-specific options`,
+        description: `The default quality. This is overridden by any format-specific options`,
       },
       jpgOptions: {
         type: JPGOptionsType,
@@ -492,6 +493,10 @@ const imageNodeType = ({
       webpOptions: {
         type: WebPOptionsType,
         description: `Options to pass to sharp when generating WebP images.`,
+      },
+      avifOptions: {
+        type: AVIFOptionsType,
+        description: `Options to pass to sharp when generating AVIF images.`,
       },
       transformOptions: {
         type: TransformOptionsType,